Blow up pipelines.py
