+++ locations/pipelines/assert_url_scheme.py
@@ -1,0 +1,7 @@
+class AssertURLSchemePipeline:
+    def process_item(self, item, spider):
+        if item.get("image"):
+            if item["image"].startswith("//"):
+                item["image"] = "https:" + item["image"]
+
+        return item
