+++ locations/pipelines/duplicates.py
@@ -1,0 +1,17 @@
+from scrapy.exceptions import DropItem
+
+
+class DuplicatesPipeline:
+    def __init__(self):
+        self.ids_seen = set()
+
+    def process_item(self, item, spider):
+        if hasattr(spider, "no_refs") and getattr(spider, "no_refs"):
+            return item
+
+        ref = (spider.name, item["ref"])
+        if ref in self.ids_seen:
+            raise DropItem("Duplicate item found: %s" % item)
+        else:
+            self.ids_seen.add(ref)
+            return item
